pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lohi_splitter"
version = "1.1"
authors = [
  { name="Simon Steshin", email="simon.steshin@gmail.com" },
]
description = "Train-test splitter for Lead Optimization and Hit Identification tasks."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mip==1.15.0",
    "tqdm",
    "rdkit-pypi>=2022.09.1",
    "networkx",
    "numpy",
]

[project.urls]
"Homepage" = "https://github.com/SteshinSS/lohi_splitter"
"Bug Tracker" = "https://github.com/SteshinSS/lohi_splitter/issues"